split azrest into separate package
