+++ llamazure/rbac/roles.py
@@ -1,36 +1,12 @@
 """Azure Role Definitions and Assignments"""
 from __future__ import annotations
 
-from typing import Any, Dict, Sequence
-from urllib.parse import urljoin
+from typing import Dict, Sequence
 
-import requests
+from llamazure.azrest.azrest import AzRest
 
 RoleDefT = Dict
 RoleAsnT = Dict
-
-
-class AzRest:
-	def __init__(self, token, session: requests.Session, base_url: str = "https://management.azure.com/"):
-		self.token = token
-		self.session = session
-
-		self.base_url = base_url
-
-	@classmethod
-	def from_credential(cls, credential) -> AzRest:
-		"""Create from an Azure credential"""
-		token = credential.get_token("https://management.azure.com//.default")
-		session = requests.Session()
-		return cls(token, session)
-
-	def call(self, req: requests.Request) -> Any:
-		"""Make the request to Azure"""
-		req.headers["Authorization"] = f"Bearer {self.token.token}"  # TODO: push down into self.session
-		return self.session.send(req.prepare()).json()["value"]  # TODO: write yet another fun interface to Azure
-
-	def get(self, slug: str) -> Any:
-		return self.call(requests.Request("GET", urljoin(self.base_url, slug)))
 
 
 class RoleDefs:
