+++ pants.toml
@@ -1,5 +1,5 @@
 [GLOBAL]
-pants_version = "2.16.0rc2"
+pants_version = "2.17.0rc2"
 backend_packages = [
 	"pants.backend.python",
 	"pants.backend.experimental.python",
